ReacTran-package(ReacTran)	R Documentation

Reactive transport modelling in 1D or 2-D

Description

R-package ReacTran contains routines that enable the development of one-dimensional and two-dimensional reactive transport models in aquatic systems (rivers, lakes) as well porous media (floc aggregates, sediments).

It contains:

Functions to setup a finite-difference grid (1D or 2D)
Functions to attach parameters and properties to the grid (1D or 2D)
Functions to calculate the advective-diffusive transport term over the grid (1D or 2D)
Utility functions

Details

Package: 	 ReacTran

Type: 	 Package

Version: 	 1.0

Date: 	 2009-02-02

License: 	 GNU Public License 2 or above

Author(s)

Karline Soetaert (Maintainer)

Filip Meysman

Examples


## Not run: 
## show examples (see respective help pages for details)
example(tran.1D)
example(tran.2D)

## open the directory with documents
browseURL(paste(system.file(package="ReacTran"), "/doc", sep=""))

## End(Not run)

[Package ReacTran version 1.1 Index]