setup.grid.1D(ReacTran)	R Documentation

Creation of a one-dimensional finite difference grid

Description

Subdivides the one-dimensional model domain into one or more zones that are each sub-divided into grid cells. The resulting grid structure can be used in the other ReacTran functions. The grid structure is characterized by the position of the middle of the grid cells (x.mid) and the position of the interfaces between cells (x.int). Distances are calculated between the interfaces (dx), i.e. the thickness of the grid cells. An auxiliary set of distances is calculated (dx.aux) between the points where the concentrations are specified (at the centers of the grid cell and the two external interfaces). A more complex grid consisting of multiple zones can be constructed when using vectors as arguments. In each zone, one can control the grid resolution near the upstream and downstream boundary. The grid resolution at the upstream interface changes according to the formula dx[i+1] = min(max.dx.1,p.dx.1*dx[i]). A similar formula controls the resolution at the downstream interface. This allows refinement of the grid near the interfaces.

Usage

setup.grid.1D(x.up=0, x.down=NULL, L=NULL, N=NULL, dx.1=NULL,
 p.dx.1=rep(1,length(L)), max.dx.1=10*dx.1, dx.N=NULL,
 p.dx.N=rep(1,length(L)), max.dx.N=10*dx.N)

## S3 method for class 'grid.1D':
plot(x, ...)

                     

Arguments

x.up 	 position of the upstream interface; one value

x.down 	 position of the endpoint of each zone; one value when the model domain covers only one zone (x.down = position of downstream interface), or a vector of length M when the model domain is divided into M zones (x.down[M] = position of downstream interface)

L 	 thickness of zones; one value (model domain = one zone) or a vector of length M (model domain = M zones)

N 	 number of grid cells within a zone; one value or a vector of length M

dx.1 	 size of the first grid cell in a zone; one value or a vector of length M

p.dx.1 	 factor controlling the increase in grid cell size near the upstream boundary; one value or a vector of length N. The default value is 1 (constant grid cell size)

max.dx.1 	 maximum grid cell size in the upstream half of the zone; one value or a vector of length M

dx.N 	 size of the last grid cell in a zone; one value or a vector of length M

p.dx.N 	 factor controlling the increase in grid cell size near the downstream boundary; one value or a vector of length N. The default value is 1 (constant grid cell size)

max.dx.N 	 maximum grid cell size in the downstream half of the zone; one value or a vector of length M

x 	 the object of class grid.1D that needs plotting

... 	 additional arguments passed to the function plot

Value

a list of type grid.1D containing:

N 	 the total number of grid cells

x.up 	 position of the upstream interface; one value

x.down 	 position of the downstream interface; one value

x.mid 	 position of the middle of the grid cells; vector of length N

x.int 	 position of the interfaces of the grid cells; vector of length N+1

dx 	 distance between adjacent cell interfaces (thickness of grid cells), vector of length N

dx.aux 	 auxiliary vector containing the distance between adjacent cell centers; at the upper and lower boundary calculated as (x[1]-x.up) and (x.down-x[N]) respectively; vector of length N+1

Author(s)

Filip Meysman <f.meysman@nioo.knaw.nl>, Karline Soetaert <k.soetaert@nioo.knaw.nl>

Examples


# 1D Grid: one zone, constant resolution 
(GR <- setup.grid.1D(x.up=0,L=10,N=10))
plot(GR)

# 1D Grid: one zone, constant resolution, origin not zero 
(GR<-setup.grid.1D(x.up=5,x.down=10,N=10))
plot(GR)

# 1D Grid: one zone, higher resolution near the upstream interface 
(GR<-setup.grid.1D(x.up=0,x.down=10,dx.1=0.1,p.dx.1=1.1))
plot(GR)

# 1D Grid: one zone, higher resolution near the upstream
# and downstream interface 
GR<-setup.grid.1D(x.up=0,x.down=10,
dx.1=0.1,p.dx.1=1.1,dx.N=0.1,p.dx.N=1.1)
plot(GR)

# 1D Grid: two zones, higher resolution near the upstream
# and downstream interface 
(GR<-setup.grid.1D(x.up=0,L=c(5,5),dx.1=c(0.2,0.2),p.dx.1=c(1.1,1.1),
                   dx.N=c(0.2,0.2),p.dx.N=c(1.1,1.1)))
plot(GR)

# 1D Grid: two zones, higher resolution near the upstream
# and downstream interface 
# the number of grid cells in each zone is imposed via N  
(GR <- setup.grid.1D(x.up=0,L=c(5,5),N=c(20,10),dx.1=c(0.2,0.2),
             p.dx.1=c(1.1,1.1),dx.N=c(0.2,0.2),p.dx.N=c(1.1,1.1)))
plot(GR)

[Package ReacTran version 1.1 Index]