setup.grid.2D(ReacTran)	R Documentation

Creation of a two-dimensional finite difference grid

Description

Creates a finite difference grid over a rectangular two-dimensional model domain starting from two separate one-dimensional grids (as created by setup.grid.1D).

Usage

setup.grid.2D(x.grid=NULL, y.grid=NULL)

Arguments

x.grid 	 list containing the one-dimensional grid in the vertical direction - see setup.grid.1D for the structure of the list

y.grid 	 list containing the one-dimensional grid in the horizontal direction - see setup.grid.1D for the structure of the list

Value

a list of type grid.2D containing:

x.up 	 vertical position of the upper interface; one value

x.down 	 vertical position of the lower interface; one value

x.mid 	 vertical position of the middle of the grid cells; vector of length x.N

x.int 	 vertical position of the horizontal interfaces of the grid cells; vector of length x.N+1

dx 	 distance between adjacent cell interfaces (thickness of grid cells); vector of length x.N

dx.aux 	 auxiliary vector containing the distance between adjacent cell centers; at the upper and lower boundary calculated as (x[1]-x.up) and (x.down-x[x.N]) respectively; vector of length x.N+1

x.N 	 total number of grid cells in the vertical direction; one value

y.left 	 horizontal position of the left interface; one value

y.right 	 horizontal position of the right interface; one value

y.mid 	 horizontal position of the middle of the grid cells; vector of length y.N

y.int 	 horizontal position of the vertical interfaces of the grid cells; vector of length y.N+1

dy 	 distance between adjacent cell interfaces (thickness of grid cells); vector of length y.N

dy.aux 	 auxiliary vector containing the distance between adjacent cell centers; at the left and right boundary calculated as (y[1]-y.left) and (y.right-y[y.N]) respectively; vector of length y.N+1

y.N 	 total number of grid cells in the horizontal direction; one value

Author(s)

Filip Meysman <f.meysman@nioo.knaw.nl>, Karline Soetaert <k.soetaert@nioo.knaw.nl>

Examples


# test of the setup.grid.2D functionality
x.grid <- setup.grid.1D(x.up=0,L=10,N=5)
y.grid <- setup.grid.1D(x.up=0,L=20,N=10)
(grid2D <- setup.grid.2D(x.grid,y.grid))

[Package ReacTran version 1.1 Index]