setup.prop.2D(ReacTran)	R Documentation

Attaches a property to a two-dimensional grid

Description

Calculates the value of a given property at the middle of grid cells (mid) and at the interfaces of the grid cells (int). Two possibilities are available: either specifying a mathematical function (func) that describes the spatial dependency of the property at hand, or asssuming a constant value (value). For example, on a sediment model, the routine can be used to specify the porosity, the mixing intensity or other parameters over the grid of the reactangular sediment domain.

Usage

setup.prop.2D(func = NULL, value = NULL, grid,...)

Arguments

func 	 Function that describes the spatial dependency

value 	 Constant value given to the property

grid 	 List specifying the 2D grid characteristics, see setup.grid.2D for details on the structure of this list

... 	 Additional arguments that are passed on to func

Value

A list of type prop.2D containing:

mid 	 property value at the middle of the grid cells; NxM matrix

x.int 	 property value at the horizontal interfaces of the grid cells; (N+1)xM matrix

y.int 	 property value at the vertical interfaces of the grid cells; Nx(M+1) matrix

Author(s)

Filip Meysman <f.meysman@nioo.knaw.nl>, Karline Soetaert <k.soetaert@nioo.knaw.nl>

Examples


# Inverse quadratic function 
inv.quad <- function(x,y,a=NULL,b=NULL)
return(1/((x-a)^2+(y-b)^2))

# Construction of the 2D grid 
x.grid <- setup.grid.1D(x.up=0,L=10,N=10)
y.grid <- setup.grid.1D(x.up=0,L=10,N=10)
grid2D <- setup.grid.2D(x.grid,y.grid)

# Attaching the inverse quadratic function to the 2D grid 
setup.prop.2D(func=inv.quad,grid=grid2D,a=5,b=5)


[Package ReacTran version 1.1 Index]