tran.2D(ReacTran)	R Documentation

General two-dimensional advective-diffusive transport

Description

Estimates the transport term (i.e. the rate of change of the concentration due to diffusion and advection) in a two-dimensional rectangular model domain.

Usage

tran.2D(C, C.up=C[1,], C.down=C[nrow(C),],
C.left=C[,1], C.right=C[,ncol(C)], 
flux.up=NA, flux.down=NA, flux.left=NA, flux.right=NA,
a.bl.up=NULL, C.bl.up=NULL, a.bl.down=NULL, C.bl.down=NULL,
a.bl.left=NULL, C.bl.left=NULL, a.bl.right=NULL, C.bl.right=NULL,
D.x=NULL, D.y=D.x, D.grid=NULL, 
v.x=0, v.y=v.x, v.grid=NULL, 
AFDW.x=1, AFDW.y=AFDW.x, AFDW.grid=NULL, 
VF.x=1, VF.y=VF.x, VF.grid=NULL,
dx=NULL, dy=NULL, grid=NULL,
full.check = FALSE, full.output = FALSE)

Arguments

C 	 Concentration, expressed per unit volume, defined at the centre of each grid cell; NxM matrix [M/L3]

C.up 	 Concentration at upstream boundary; vector of length M [M/L3]

C.down 	 Concentration at downstream boundary; vector of length M [M/L3]

C.left 	 Concentration at left boundary; vector of length N [M/L3]

C.right 	 Concentration at right boundary; vector of length N [M/L3]

flux.up 	 Flux across the upstream boundary, positive = INTO model domain; vector of length M [M/L2/T]

flux.down 	 Flux across the downstream boundary, positive = OUT of model domain; vector of length M [M/L2/T]

flux.left 	 Flux across the left boundary, positive = INTO model domain; vector of length N [M/L2/T]

flux.right 	 Flux across the right boundary, positive = OUT of model domain; vector of length N [M/L2/T]

a.bl.up 	 Transfer coefficient across the upstream boundary layer. Flux=a.bl.up*(C.bl.up-C[1,]). One value [L/T]

C.bl.up 	 Concentration at the upstream boundary layer; vector of length M [M/L3]

a.bl.down 	 Transfer coefficient across the downstream boundary layer; Flux=a.bl.down*(C[N,]-C.bl.down). One value [L/T]

C.bl.down 	 Concentration at the downstream boundary layer; vector of length M [M/L3]

a.bl.left 	 Transfer coefficient across the left boundary layer. Flux=a.bl.left*(C.bl.left-C[,1]). One value [L/T]

C.bl.left 	 Concentration at the left boundary layer; vector of length N [M/L3]

a.bl.right 	 Transfer coefficient across the right boundary layer; Flux=a.bl.right*(C[,M]-C.bl.right). One value [L/T]

C.bl.right 	 Concentration at the right boundary layer; vector of length N [M/L3]

D.x 	 diffusion coefficient in x-direction, defined on grid cell interfaces. One value or an (N+1)x(M) matrix [L2/T]

D.y 	 diffusion coefficient in y-direction, defined on grid cell interfaces. One value or an (N)x(M+1) matrix [L2/T]

D.grid 	 diffusion coefficient packaged as a grid list; the list contains at least the elements x.int and y.int (see setup.prop.2D) [L2/T]

v.x 	 advective velocity in the x-direction, defined on grid cell interfaces. Can be positive (downstream flow) or negative (upstream flow). One value or an (N+1)x(M) matrix [L/T]

v.y 	 advective velocity in the y-direction, defined on grid cell interfaces. Can be positive (downstream flow) or negative (upstream flow). One value or an (N)x(M+1) matrix [L/T]

v.grid 	 advective velocity packaged as a grid list; the list contains at least the elements x.int and y.int (see setup.prop.2D) [L/T]

AFDW.x 	 weight used in the finite difference scheme for advection in the x-direction, defined on grid cell interfaces; backward = 1, centred = 0.5, forward = 0; default is backward. One value or an (N+1)x(M) matrix [-]

AFDW.y 	 weight used in the finite difference scheme for advection in the y-direction, defined on grid cell interfaces; backward = 1, centred = 0.5, forward = 0; default is backward. One value or an (N)x(M+1) matrix [-]

AFDW.grid 	 AFDW coefficients packaged as a grid list; the list contains at least the elements x.int and y.int (see setup.prop.2D) [-]

VF.x 	 Volume fraction at the grid cell interfaces in the x-direction. One value or an (N+1)x(M) matrix [-]

VF.y 	 Volume fraction at the grid cell interfaces in the y-direction. One value or an (N)x(M+1) matrix [-]

VF.grid 	 Volume fraction packaged as a grid list; the list contains at least the elements x.int, y.int, 'x.mid' and 'y.mid' (see setup.prop.2D) [-]

dx 	 distance between adjacent cell interfaces in the x-direction (thickness of grid cells). One value or vector of length N [L]

dy 	 distance between adjacent cell interfaces in the y-direction (thickness of grid cells). One value or vector of length M [L]

grid 	 discretization grid, a list containing at least elements dx, dx.aux, dy, dy.aux (see setup.grid.2D) [L]

full.check 	 logical flag enabling a full check of the consistency of the arguments (default = FALSE; TRUE slows down execution by 50 percent)

full.output 	 logical flag enabling a full return of the output (default = FALSE; TRUE slows down execution by 20 percent)

Details

The boundary conditions are either

(1) zero-gradient
(2) fixed concentration
(3) convective boundary layer
(4) fixed flux

This is also the order of priority. The zero gradient is the default, the fixed flux overrules all other.

Value

when (full.output = FALSE) then a list containing:

dC 	 The rate of change of the concentration C due to transport, defined in the centre of each grid cell [M/L3/T]

dC 	 The rate of change of the concentration C due to transport, defined in the centre of each grid cell [M/L3/T]

C.up 	 Concentration at the upstream interface. One value [M/L3]

C.down 	 Concentration at the downstream interface. One value [M/L3]

adv.flux 	 Advective flux across at the interface of each grid cell. A vector of length N+1 [M/L2/T]

dif.flux 	 Diffusive flux across at the interface of each grid cell. A vector of length N+1 [M/L2/T]

flux 	 Total flux across at the interface of each grid cell. A vector of length N+1 [M/L2/T]

flux.up 	 Flux across the upstream boundary, positive = INTO model domain. One value [M/L2/T]

flux.down 	 Flux across the downstream boundary, positive = OUT of model domain. One value [M/L2/T]

Author(s)

Filip Meysman <f.meysman@nioo.knaw.nl>, Karline Soetaert <k.soetaert@nioo.knaw.nl>

References

Soetaert and Herman, a guide to ecological modelling - using R as a simulation platform, 2008. Springer

Examples


# Parameters
F        <- 100             # input flux [micromol cm-2 yr-1]
por      <- 0.8             # constant porosity
D        <- 400             # mixing coefficient [cm2 yr-1]
v        <- 1               # advective velocity [cm yr-1]

# Grid definition
x.N <- 4   # number of cells in x-direction
y.N <- 6   # number of cells in y-direction
x.L <- 8   # domain size x-direction [cm]
y.L <- 24  # domain size y-direction [cm]
dx       <- x.L/x.N             # cell size x-direction [cm]
dy       <- y.L/y.N             # cell size y-direction [cm]
 
# Intial conditions 
C <- matrix(nrow=x.N,ncol=y.N,data=0,byrow=FALSE)

# Boundary conditions: fixed concentration  
C.up=rep(1,times=y.N)
C.down=rep(0,times=y.N)
C.left=rep(1,times=x.N)
C.right=rep(0,times=x.N)

# Only diffusion 
tran.2D(full.output=TRUE,C=C,D.x=D,D.y=D,v.x=0,
v.y=0,VF.x=por,VF.y=por,dx=dx,dy=dy,C.up=C.up,C.down=C.down,
C.left=C.left,C.right=C.right)  

# Strong advection, backward (default), central and forward 
#finite difference schemes 
tran.2D(C=C,D.x=D,v.x=100*v,VF.x=por,dx=dx,dy=dy,
C.up=C.up,C.down=C.down,C.left=C.left,C.right=C.right)  
tran.2D(AFDW.x=0.5,C=C,D.x=D,v.x=100*v,VF.x=por,dx=dx,dy=dy,
C.up=C.up,C.down=C.down,C.left=C.left,C.right=C.right)  
tran.2D(AFDW.x=0,C=C,D.x=D,v.x=100*v,VF.x=por,dx=dx,dy=dy,
C.up=C.up,C.down=C.down,C.left=C.left,C.right=C.right)  

# Boundary conditions: fixed fluxes 

flux.up=rep(200,times=y.N)
flux.down=rep(-200,times=y.N)
flux.left=rep(200,times=x.N)
flux.right=rep(-200,times=x.N)
tran.2D(C=C,D.x=D,v.x=0,VF.x=por,dx=dx,dy=dy,
flux.up=flux.up,flux.down=flux.down,
flux.left=flux.left,flux.right=flux.right)  

# Boundary conditions: convective boundary layer on all sides

a.bl <- 800   # transfer coefficient
C.bl.up <- rep(1,times=(y.N)) # fixed conc at boundary layer
C.bl.left <- rep(1,times=(x.N)) # fixed conc at boundary layer
tran.2D(full.output=TRUE,C=C,D.x=D,v.x=0,VF.x=por,
dx=dx,dy=dy,C.bl.up=C.bl.up,a.bl.up=a.bl,C.bl.down=C.bl.up,
a.bl.down=a.bl,C.bl.left=C.bl.left,a.bl.left=a.bl,
C.bl.right=C.bl.left,a.bl.right=a.bl)  

# Runtime test with and without argument checking

n.iterate <-1000

test1 <- function()
{
for (i in 1:n.iterate )
ST<-tran.2D(full.check=TRUE,C=C,D.x=D,v.x=0,VF.x=por,
dx=dx,dy=dy,C.bl.up=C.bl.up,a.bl.up=a.bl,C.down=C.down)  
} 
system.time(test1())

test2 <- function()
{
for (i in 1:n.iterate )
ST<-tran.2D(full.output=TRUE,C=C,D.x=D,v.x=0,VF.x=por,
dx=dx,dy=dy,C.bl.up=C.bl.up,a.bl.up=a.bl,C.down=C.down)  
} 
system.time(test2())

test3 <- function()
{
for (i in 1:n.iterate )
ST<-tran.2D(full.output=TRUE,full.check=TRUE,C=C,D.x=D,v.x=0,
VF.x=por,dx=dx,dy=dy,C.bl.up=C.bl.up,a.bl.up=a.bl,C.down=C.down)  
} 
system.time(test3())


[Package ReacTran version 1.1 Index]