tran.volume.1D(ReacTran)	R Documentation

1-D volumetric advective-diffusive transport in an aquatic system

Description

Estimates the volumetric transport term (i.e. the rate of change of the concentration due to diffusion and advection) in a one-dimensional model of an aquatic system (river, estuary). Volumetric transport implies the use of flows (mass per unit of time) rather than fluxes (mass per unit of area per unit of time) as is done in tran.1D. The tran.volume.1D routine is particularly suited for modelling channels (like rivers, estuaries) where the cross-sectional area changes, but where this area chnage is not explicitly modelled as such. Another difference withy tran.1D is that the present routine also allows the input of lateral water flow.

Usage

tran.volume.1D(C, C.up=C[1], C.down=C[length(C)],
C.lat=0, C.lat.grid=list(mid=rep(C.lat,length.out=length(C))), 
F.up=NA, F.down=NA, 
F.lat=NA, F.lat.grid=list(mid=rep(F.lat,length.out=length(C))),
Disp=NULL, Disp.grid=list(int=rep(Disp,length.out=(length(C)+1))),
flow = 0, flow.grid=list(int=rep(flow,length.out=(length(C)+1))),
flow.lat=0, flow.lat.grid=list(mid=rep(flow.lat,length.out=length(C))),
V=NULL,V.grid=list(mid=rep(V,length.out=length(C))))

Arguments

C 	 Tracer concentration, defined at the centre of the grid cells. A vector of length N [M/L3]

C.up 	 Tracer concentration at the upstream interface. One value [M/L3]

C.down 	 Tracer concentration at downstream interface. One value [M/L3]

C.lat 	 Tracer concentration in the lateral input, defined at grid cell centres. One value or a vector of length N [M/L3]

C.lat.grid 	 Tracer concentration in lateral input defined as grid property

F.up 	 Total tracer input at the upstream interface. One value [M/T]

F.down 	 Total tracer input at downstream interface. One value [M/T]

F.lat 	 Total lateral tracer input, defined at grid cell centres. One value or a vector of length N [M/T]

F.lat.grid 	 Total lateral tracer input defined as grid property

Disp 	 BULK dispersion coefficient, defined on grid cell interfaces. One value or a vector of length N+1, [L3/T]

Disp.grid 	 BULK dispersion coefficient defined as grid property

flow 	 Water flow rate, defined on grid cell interfaces. One value or a vector of length N+1, [L3/T]

flow.grid 	 Water flow rate defined as grid property

flow.lat 	 Lateral water flow rate into each volume box, defined at grid cell centres. One value or a vector of length N, [L3/T]

flow.lat.grid 	 Lateral water flow rate defined as grid property

V 	 Grid cell volume, defined at grid cell centres. One value or a vector of length N [L3]

V.grid 	 Grid cell volume defined as grid property

Details

The boundary conditions are of type

1. zero-gradient (default)
2. fixed concentration
3. fixed input

The bulk dispersion coefficient (Disp) and the flow rate (flow) can be either one value or a vector of length N+1, defined at all grid cell interfaces, including upstream and downstream boundary.

The spatial discretisation is given by the volume of each box (V), which can be one value or a vector of length N+1, defined at the centre of each grid cell. The water flow is mass conservative. Over each volume box, the routine calculates internally the downstream outflow of water in terms of the upstream inflow and teh lateral inflow.

Value

dC 	 The rate of change of the concentration C due to transport, defined in the centre of each grid cell [M/L3/T]

F 	 Mass flow across at the interface of each grid cell. A vector of length N+1 [M/T]

F.up 	 Mass flow across the upstream boundary, positive = INTO model domain. One value [M/T]

F.down 	 Mass flow across the downstream boundary, positive = OUT of model domain. One value [M/T]

F.lat 	 Lateral mass input per volume box, positive = INTO model domain. A vector of length N [M/T]

Author(s)

Filip Meysman <f.meysman@nioo.knaw.nl>, Karline Soetaert <k.soetaert@nioo.knaw.nl>

References

Soetaert and Herman (2008) A practical guide to ecological modelling - using R as a simulation platform. Springer.

See Also

tran.1D

Examples



##################################################################
#  EXAMPLE : organic carbon (OC) decay in a widening river       #
##################################################################

# Two scenarios are simulated: the baseline includes only input 
# of organic matter upstream. The second scenario simulates the 
# input of an important side river half way the estuary.  

#====================#
# Model formulation  #
#====================#

river.model <- function (t=0,OC,pars=NULL)
{
tran <- tran.volume.1D(C=OC,F.up=F.OC,F.lat=F.lat,Disp=Disp,
flow=flow,V=Volume)$dC 
reac <- - k*OC
return(list(dCdt = tran + reac))
}

#======================#
# Parameter definition #
#======================#

# Initialising morphology estuary: 

nbox          <- 500     # number of grid cells
lengthEstuary <- 100000  # length of estuary [m]
BoxLength     <- lengthEstuary/nbox # [m]
Distance      <- seq(BoxLength/2, by=BoxLength, len=nbox) # [m]

# Cross sectional area: sigmoid function of estuarine distance [m2]
CrossArea <- 4000 + 72000 * Distance^5 /(Distance^5+50000^5)

# Volume of boxes                          (m3)
Volume  <- CrossArea*BoxLength

# Transport coefficients
Disp    <- 1000   # m3/s, bulk dispersion coefficient
flow    <- 180    # m3/s, mean river flow

F.OC    <- 180               # input organic carbon [mol s-1]
F.lat.0 <- F.OC              # lateral input organic carbon [mol s-1]

k       <- 10/(365*24*3600)  # decay constant organic carbon [s-1]

#====================#
# Model solution     #
#====================#

F.lat <- rep(0,length.out=nbox)
Conc1 <- steady.band(runif(nbox),fun=river.model,nspec=1)$y   
F.lat <- F.lat.0*dnorm(x=Distance/lengthEstuary,
mean = Distance[nbox/2]/lengthEstuary, sd = 1/20, log = FALSE)/nbox 
Conc2 <- steady.band(runif(nbox),fun=river.model,nspec=1)$y   

#====================#
# Plotting output    #
#====================#

matplot(Distance/1000,cbind(Conc1,Conc2),lwd=2,
main="Organic carbon decay in an estuary",xlab="distance [km]",
ylab="OC Concentration [mM]",
type="l")  
legend ("topright",lty=1,col=c("black","red"),
        c("baseline","with lateral input"))

[Package ReacTran version 1.1 Index]